% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_files_to_pdf.R
\name{convert_files_to_pdf}
\alias{convert_files_to_pdf}
\title{Convert Word & Excel Files to PDFs using 'VBScript'}
\usage{
convert_files_to_pdf(folder_path)
}
\arguments{
\item{folder_path}{The directory containing Word and Excel files}
}
\value{
No return value, called for side effects.
}
\description{
Convert Word & Excel Files to PDFs using 'VBScript'
}
