% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_html_htm_to_pdf.R
\name{convert_html_to_pdf}
\alias{convert_html_to_pdf}
\title{Convert HTML/HTM Files to PDF
Automatically installs 'PhantomJS' if missing.}
\usage{
convert_html_to_pdf(folder_path)
}
\arguments{
\item{folder_path}{The directory where ZIP files are extracted.}
}
\value{
No return value, called for side effects.
}
\description{
Convert HTML/HTM Files to PDF
Automatically installs 'PhantomJS' if missing.
}
