% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_zip_files.R
\name{extract_zip_files}
\alias{extract_zip_files}
\title{Extract ZIP Files in a Given Folder}
\usage{
extract_zip_files(folder_path)
}
\arguments{
\item{folder_path}{The directory containing ZIP files}
}
\value{
No return value, called for side effects.
}
\description{
Extract ZIP Files in a Given Folder
}
