% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_pdfs.R
\name{merge_pdfs}
\alias{merge_pdfs}
\title{Merge PDFs in Alphabetical Order}
\usage{
merge_pdfs(folder_path)
}
\arguments{
\item{folder_path}{The directory containing sub-folders with PDF files}
}
\value{
Character vector with the path of the merged PDF file.
}
\description{
Merge PDFs in Alphabetical Order
}
