% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Full Process: Extract, Convert, and Merge PDFs}
\usage{
run_app(folder_path = NULL)
}
\arguments{
\item{folder_path}{The directory containing files for processing. If NULL, the user is prompted to select a directory interactively.}
}
\value{
No return value, called for side effects.
}
\description{
Run the Full Process: Extract, Convert, and Merge PDFs
}
